+++ src/Products/PluginIndexes/CompositeIndex/tests/testCompositeIndex.py
@@ -275,7 +275,7 @@ class CompositeIndexPerformanceTest(CompositeIndexTestMixin,
                 # if length of result is greater than zero composite
                 # search must be roughly faster than default search
                 if res1 and res2:
-                    assert 0.9 * duration2 < duration1, (duration2, duration1)
+                    assert 0.5 * duration2 < duration1, (duration2, duration1)
 
             # is result identical?
             self.assertEqual(len(res1), len(res2), '%s != %s for %s' %